+++ test/unit/panel_controller_test.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:panel_layout/src/panel_controller.dart';
 import 'package:panel_layout/src/panel_data.dart';